moved Sender Attributes to a separate class
