+++ SpeckleGrasshopper/BaseComponents/GhSenderCoreClient.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Diagnostics;
-using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -12,13 +11,12 @@ using System.Threading.Tasks;
 using System.Timers;
 using System.Windows.Forms;
 using GH_IO.Serialization;
-using Grasshopper.GUI;
 using Grasshopper.GUI.Base;
-using Grasshopper.GUI.Canvas;
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Parameters;
 using Grasshopper.Kernel.Special;
 using SpeckleCore;
+using SpeckleGrasshopper.Attributes;
 using SpeckleGrasshopper.Properties;
 
 namespace SpeckleGrasshopper
@@ -1163,77 +1161,6 @@ namespace SpeckleGrasshopper
     {
       get { return new Guid("{e66e6873-ddcd-4089-93ff-75ae09f8ada3}"); }
     }
-  }
-
-  public class GhSenderClientAttributes : Grasshopper.Kernel.Attributes.GH_ComponentAttributes
-  {
-    private GhSenderClient Base;
-    private Rectangle BaseRectangle;
-    private Rectangle StreamIdBounds;
-    private Rectangle StreamNameBounds;
-    private Rectangle PushStreamButtonRectangle;
-
-    public GhSenderClientAttributes(GhSenderClient component) : base(component)
-    {
-      Base = component;
-    }
-
-    protected override void Layout()
-    {
-      base.Layout();
-      BaseRectangle = GH_Convert.ToRectangle(Bounds);
-      StreamIdBounds = new Rectangle((int)(BaseRectangle.X + (BaseRectangle.Width - 120) * 0.5), BaseRectangle.Y - 25, 120, 20);
-      StreamNameBounds = new Rectangle(StreamIdBounds.X, BaseRectangle.Y - 50, 120, 20);
-
-      PushStreamButtonRectangle = new Rectangle((int)(BaseRectangle.X + (BaseRectangle.Width - 30) * 0.5), BaseRectangle.Y + BaseRectangle.Height, 30, 30);
-
-      if (Base.ManualMode)
-      {
-        Rectangle newBaseRectangle = new Rectangle(BaseRectangle.X, BaseRectangle.Y, BaseRectangle.Width, BaseRectangle.Height + 33);
-        Bounds = newBaseRectangle;
-      }
-    }
-
-    protected override void Render(GH_Canvas canvas, Graphics graphics, GH_CanvasChannel channel)
-    {
-      base.Render(canvas, graphics, channel);
-
-      if (channel == GH_CanvasChannel.Objects)
-      {
-        GH_PaletteStyle myStyle = new GH_PaletteStyle(System.Drawing.ColorTranslator.FromHtml(Base.EnableRemoteControl ? "#147DE9" : "#B3B3B3"), System.Drawing.ColorTranslator.FromHtml("#FFFFFF"), System.Drawing.ColorTranslator.FromHtml(Base.EnableRemoteControl ? "#ffffff" : "#4C4C4C"));
-
-        GH_PaletteStyle myTransparentStyle = new GH_PaletteStyle(System.Drawing.Color.FromArgb(0, 0, 0, 0));
-
-        var streamIdCapsule = GH_Capsule.CreateTextCapsule(box: StreamIdBounds, textbox: StreamIdBounds, palette: Base.EnableRemoteControl ? GH_Palette.Black : GH_Palette.Transparent, text: Base.EnableRemoteControl ? "Remote Controller" : "ID: " + (Base.Client != null ? Base.Client.StreamId : "error"), highlight: 0, radius: 5);
-        streamIdCapsule.Render(graphics, myStyle);
-        streamIdCapsule.Dispose();
-
-        var streamNameCapsule = GH_Capsule.CreateTextCapsule(box: StreamNameBounds, textbox: StreamNameBounds, palette: GH_Palette.Black, text: "(S) " + Base.NickName, highlight: 0, radius: 5);
-        streamNameCapsule.Render(graphics, myStyle);
-        streamNameCapsule.Dispose();
-
-        if (Base.ManualMode)
-        {
-          var pushStreamButton = GH_Capsule.CreateCapsule(PushStreamButtonRectangle, GH_Palette.Pink, 2, 0);
-          pushStreamButton.Render(graphics, true ? Properties.Resources.play25px : Properties.Resources.pause25px, myTransparentStyle);
-        }
-      }
-    }
-
-    public override GH_ObjectResponse RespondToMouseDown(GH_Canvas sender, GH_CanvasMouseEvent e)
-    {
-      if (e.Button == System.Windows.Forms.MouseButtons.Left)
-      {
-        if (((RectangleF)PushStreamButtonRectangle).Contains(e.CanvasLocation))
-        {
-          Base.ManualUpdate();
-          //Base.ExpireSolution( true );
-          return GH_ObjectResponse.Handled;
-        }
-      }
-      return base.RespondToMouseDown(sender, e);
-    }
-
   }
 }
