+++ SpeckleGrasshopper/SpeckleGrasshopper.csproj
@@ -70,6 +70,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BaseComponents\GhReceiverCoreClient.cs" />
+    <Compile Include="Attributes\GhSenderClientAttributes.cs" />
     <Compile Include="BaseComponents\GhSenderCoreClient.cs" />
     <Compile Include="Management\CreateAccount.cs" />
     <Compile Include="UserDataUtils\GetApplicationId.cs" />
